test: split preflight argument tests into test_preflight_arguments
Relocate the canonical preflight argument-composition tests
(test_preflight_argument_set_invariants and test_preflight_argument_sets_snapshot)
plus their _dir_name Hypothesis strategy out of test_preflight_checks and into
the existing test_preflight_arguments module, migrating the two composed-argument
snapshots to test_preflight_arguments.ambr. test_preflight_checks now focuses on
the remaining pre-flight-checks coverage, and its now-unused hypothesis/string
imports are dropped.

Addresses the reviewer's request to keep the argument-composition tests in a
dedicated module. (The originally cited "module length limit" does not itself
apply — the file was 438 lines against the 800-line max, and too-many-lines is
disabled — but the cohesion split is a reasonable, low-risk reorganisation.)
