+++ tests/unit/publish/test_preflight_checks.py
@@ -3,13 +3,10 @@
 from __future__ import annotations
 
 import collections.abc as cabc
-import string
 import typing as typ
 from pathlib import Path
 
-import hypothesis.strategies as st
 import pytest
-from hypothesis import HealthCheck, given, settings
 
 from lading.commands import publish, publish_preflight
 
@@ -383,56 +380,3 @@ def test_verify_clean_working_tree_reports_missing_repo(
     # Lock the operator-facing missing-repository message (issue #96
     # failure-message snapshot coverage) so its wording cannot drift silently.
     assert message == snapshot()
-
-
-# ---------------------------------------------------------------------------
-# Canonical preflight argument composition (issue #96)
-# ---------------------------------------------------------------------------
-
-_dir_name = st.text(
-    alphabet=string.ascii_letters + string.digits + " -_!@#&",
-    min_size=1,
-    max_size=20,
-).filter(lambda name: name.strip(" ") == name)
-
-
-@given(unit_tests_only=st.booleans(), dir_name=_dir_name)
-@settings(max_examples=20, suppress_health_check=[HealthCheck.function_scoped_fixture])
-def test_preflight_argument_set_invariants(
-    *, unit_tests_only: bool, dir_name: str
-) -> None:
-    """Composed cargo arguments satisfy the documented invariants.
-
-    ``--workspace`` always leads, ``--target-dir`` always points at the
-    supplied directory, ``--all-targets`` appears in the check set always and
-    in the test set iff full test targets are requested, and composition is
-    deterministic.
-    """
-    target_dir = Path("/preflight") / dir_name
-
-    check_args, test_args = publish_preflight._preflight_argument_sets(
-        target_dir, unit_tests_only=unit_tests_only
-    )
-
-    for arguments in (check_args, test_args):
-        assert arguments[0] == "--workspace"
-        assert arguments[-1] == f"--target-dir={target_dir}"
-    assert "--all-targets" in check_args
-    assert ("--all-targets" in test_args) == (not unit_tests_only)
-    assert (check_args, test_args) == publish_preflight._preflight_argument_sets(
-        target_dir, unit_tests_only=unit_tests_only
-    )
-
-
-@pytest.mark.parametrize("unit_tests_only", [True, False])
-def test_preflight_argument_sets_snapshot(
-    snapshot: SnapshotAssertion,
-    *,
-    unit_tests_only: bool,
-) -> None:
-    """The composed cargo argument tuples are locked by snapshot."""
-    check_args, test_args = publish_preflight._preflight_argument_sets(
-        Path("/preflight/target"), unit_tests_only=unit_tests_only
-    )
-
-    assert snapshot == {"check": check_args, "test": test_args}
